nonlin integer

+++ nonlinearity.py
@@ -1,4 +1,4 @@
 top = 255
 
 def compensate(original):
-    return (original * original) / top
+    return int((original * original) / top)
